+++ CHANGES.rst
@@ -9,6 +9,9 @@ Changelog
   [gbastien]
 - Pinned `z3c.unconfigure` to `2.1` instead `3.0` as it breaks Plone6.0 tests.
   [gbastien]
+- Overrided `collective.quickupload` JS function `_addSelection` to fix selected
+  elements order (order was reversed).
+  [gbastien]
 
 1.2.3 (2025-03-11)
 ------------------